Booking
❇️ Booking

Create a Booking: The user can book a property for specific dates.
If the booking is approved and not canceled, then another booking with overlapping dates cannot be created
View Bookings: The user can view their active and completed bookings.
Cancel a Booking: The user can cancel a booking before a specific date.
Confirm a Booking: The landlord can confirm or cancel booking requests.